third_party/blink/web_tests/external/wpt/css/css-sizing/contain-intrinsic-size/auto-011-expected.txt

This is a testharness.js-based test.
FAIL contain:size does not prevent recording last remembered size assert_equals: Using last remembered size - clientWidth expected 20 but got 2
PASS contain:inline-size does not prevent recording last remembered inline size
Harness: the test ran to completion.
